requirements.txt
Flask==2.3.3
tensorflow-cpu==2.13.0
Pillow==10.3.0
numpy==1.26.4
